Extracted a MySqlConnection class out of the MySqlGuest class to better show its purpose in the interface. Moved the "generate_password" function out of the class and into the nova::guest::mysql namespace.
Changed Log.debug to print to std::err.

Changed apt::fix so that when "dpkg --configure -a" is called if it writes to std::out the call won't be considered a failure. Now it just assumes it worked if it eventually see EOF, which is the best it can do.

Created overloads of the "execute" and "wait_for_eof" functions of the Process class which capture stdout to a stream. This made it easier to check some things in the integration tests.

Removed MySqlGuest::disable_root.

Integration testing revealed nearly nothing in the MySqlGuest class worked which seemed to be an issue. The culprit is that the MySQL prepared statements don't work for most administrative commands. We're going to probably end up escaping stuff client side which feels wrong. I can't figure out how to call the escape string method in the normal C libs either, so for now there's a dummy method which returns the same string passed in.

Got beguiled by the notion of reusing our real Proboscis-style integration tests by wrapping the new agent code in Python. Boost Python was easy to use initially and it looked as if we'd get it for free. However as I began to wrap the various POCOs (plain old C objects) in the nova::guest::mysql namespace the wrapping process got difficult and I realized no one would appreciate the added complexity. Thankfully, even though the Python code had been making use of Pexpect and other libs the integration tests were easier to rewrite in C++ than I anticipated. However the end result isn't as readable as I'd have liked.