Use shared_ptr for objects sharing AptGuest.
I switched multiple objects who had AptGuest from a using a reference
to using boost::shared_ptr, since the fact is that object *is* shared.
In reality the old code was fine, but I felt like it might be
confusing and inconsistent to be doing things with references like that.