Don't use refs for extra backup update args.
I had used references for a struct passed to the backup update_db
function as it seemed efficient to pass in const, but the more I think
on it this feels unwise (premature optimization). It seems to work in
my VM but I'm changing it just to err on the side of caution.