Changes to sneaky-pete for modify-user-attributes
At present, when a Cloud Database user creates a username, password and hostname for a database user,
he does not have the ability to modify these user attributes.
This API enables the Cloud DB user to make changes to one or more of these user attributes (username,hostname,password).
These changes are in the sneaky-pete guest agent.

Implements: blueprint modify-userattributes