Cleaner integration of build info into executable.
This commit gets rid of the neasty "Version Header Generator" that
nobody liked and instead passes in the VERSION_STRING via the command
line. This string is whats used when the debian packages are built.
This also puts the build number into the flag as well as the
diagnostics call.