Adds back check for XtraBackup success message.
This had to be reverted earlier as it failed in real mode tests
(despite passing functional tests). The fix involved storing an
extra two characters in the CabooseChecker and then checking that
the success string was found in the trailing string, instead of
checking that the two strings were identical.