Add format and mount volume capability
If a cinder volume is created for a trove instance, in order to
use it for /var/lib/mysql (or any directory specified in the config)
it requires the agent to format and mount the volume to the specfied
path.

Formatting and mounting of volume occurs within the "prepare" message
handler.

This feature can be enabled with the 'volume_format_and_mount' flag.