Adds ResilientSender class and Conductor driver
Corrects spelling of "Resilent" to "Resilient" where applicable.
Guest Agent now uses Conductor to update status, instead of db connection.
Instead of backup_id, backup_info is passed to agent.
send_to_conductor will send a message to conductor on the host.
With this commit the host database requirement is not yet removed.