Fixing memory leak.
The BackupManager was originally created on the stack
of the main method, which was ok as it could live there
forever.

With my recent changes however this method no longer
persisted, and unfortunately some Json RPC marshallers
held references (this was dumb). Changing these to be
full copies fixed everything.