Adds python script to generate status code
The Trove Service Status maps to the Sneaky Pete DatastoreStatus
type's description strings and integers. This commit introduces a
Python script that runs Trove and generates the relevant source code
to keep this sync'd up. The script also can be run as a test to see
if the source code's values have changed from what's in Trove.