Add enable_ssl message handler
This adds an enable_ssl method to the datastore apps to be implemented.
The MySQL implementation drops the key files to a location in /etc/mysql
and drops an extra my.cnf extention with ssl configs.

This also adds a message handler for the mysql message handler to accept
new ssl keys. This handler does not restart the db and can be used to
configure an instance for ssl on existing instances. It will simply
not be available until the instance is restarted.