Stopped doing some things in the Redis client
The client was looking for the renamed config command, but we ended up
not needing to do this. It also had a config class but that was only
being used in the constructor so I zapped it. Also removed some more
constants from constants.h that were only used in a few places.