Fixing bug that inhibited restores
This bug caused Maria and Percona to not initialize correctly
during a restore since an attempt was made to delete databases
that did not exist. For whatever reason that made the mysqld_safe
process invunerable to attacks from /etc/init.d/mysql stop.
This commit changes the logic to not try to delete those things
during a restore, since in theory they should already be gone.